Attempt to upgrade sqlite3 when running tests on CI
The ubuntu-latest runner for GitHub Actions (currently 22.04) uses a
version of sqlite3 which lacks support for FULL OUTER JOIN, a feature
required by one of my SQL queries. In an attempt to fix this, we will
try upgrading sqlite3 as part of the action (prior to running tests).

For more details, see:
<https://github.com/actions/runner-images/issues/9033#issuecomment-1857780470>