Revert "Attempt to upgrade sqlite3 when running tests on CI"
This reverts commit 3f4eb972a9220d377ae98e5790e1ebf2defd1a59.

This did not end up working, as Python includes its own version of
sqlite3, and even if we can jump through the hoops to upgrade Python's
version of sqlite3 (which is fraught with its own challenges and risks),
it may impact the Python version compatibility for this project. So
instead, I'm going to opt for replacing the FULL OUTER JOIN with a LEFT
JOIN, since sqlite3 has had solid support for that type of join.